Altura e empilhamento, sem andares
O jogo nao tem andares: o mapa inteiro fica numa camada de z e o relevo vem
de itens com CONST_PROP_HASHEIGHT empilhados no MESMO tile. Cada item da
pilha e um nivel, sem limite, e cada nivel empurra o personagem 8px para
cima na tela.

SERVER -- Game::internalMoveCreature

O codigo que estava la era o do Tibia: usava hasHeight para MUDAR DE ANDAR
(destPos.z-- / z++). Isso nao se aplica e foi removido.

No lugar, o JUMP passa a ser a diferenca de altura que o personagem vence de
uma vez: compara getHeightLevels() do tile atual com o do destino, e barra
se a subida passar do jump. Descer e sempre livre, como no FFTA.

Tile::getHeightLevels e novo porque hasHeight(n) nao servia: ele so responde
"tem EXATAMENTE n" e para de contar ao chegar la. Para comparar dois tiles e
preciso o total.

CLIENT -- sem teto na elevacao

m_drawElevation era uint8 e o acumulo passava por
std::min(..., MAX_ELEVATION=248) em 7 pontos. Com empilhamento sem limite o
teto nao faz sentido -- uma pilha de 32 itens de 8px ja passa de 255. O tipo
virou int e o min saiu.

Tambem corrigido o redesenho de tiles vizinhos em Tile::drawBottom: ele
somava `x * spriteSize` (grade ortogonal) e desenhava a criatura deslocada
por cima do mapa. Agora o offset e projetado no losango.

MAPA DE TESTE

Degraus de 1 a 6 niveis logo ao lado do nascimento (10..15, 8), uma pilha de
12 niveis em (12,12) para provar que nao ha limite, e uma poca de agua a
dois passos do temple.

+++ client/src/client/tile.h
@@ -179,7 +179,9 @@ private:
     std::vector<EffectPtr> m_effects; // leave this outside m_things because it has no stackpos.
     std::vector<ThingPtr> m_things;
     Position m_position;
-    uint8 m_drawElevation;
+    // int, nao uint8: o empilhamento nao tem limite, e 255 e pouco -- uma
+    // pilha de 32 itens de 8px ja passa disso.
+    int m_drawElevation;
     uint8 m_minimapColor;
     uint32 m_flags, m_houseId;
     uint16 m_speed = 0;
